refactor(release): drop tools/release sub-toolchain; use root
Per #11896 review: avoid mixing different coding principles in a
nested part of the repo. The conductor's runtime PHP code lived in a
self-contained sub-package under tools/release/ with its own composer,
phpunit, phpcs, phpstan, and rector configs. That mismatch was the
real concern, not just phpcs.

Folds the conductor into the root toolchain:

  - Deletes tools/release/{composer.json, composer.lock, phpcs.xml,
    phpstan.neon, phpunit.xml.dist, rector.php, .gitignore, vendor/}.
  - Adds justinrainbow/json-schema ^6.8 to root composer.json
    require-dev (the only direct dep that wasn't already there;
    symfony/http-client and symfony/process were).
  - Adds an OpenEMR\Release\ -> tools/release/src/ PSR-4 mapping under
    autoload-dev (not autoload). The conductor classes only run from
    CI workflows where dev deps are installed; keeping them out of
    production autoload means composer-require-checker doesn't demand
    json-schema and http-client-contracts as production deps either.
  - Moves tools/release/tests/ to tests/Tests/Isolated/Release/ with
    namespace OpenEMR\Tests\Isolated\Release so the existing root
    phpunit-isolated config picks them up.
  - Adds tools/release/{src,bin} to root phpstan.neon.dist paths so
    they are analysed by the same level-10 PHPStan run. Adjusts two
    catch (Throwable) blocks to catch RuntimeException (and
    InvalidArgumentException for the dispatcher) per the project's
    forbiddenCatchType rule.
  - Updates tools/release/bin/*.php to require the root
    vendor/autoload.php instead of the now-deleted nested one.
  - Drops the .gitignore exception for the deleted nested phpcs/phpstan
    config files.

The tools/release/ directory remains as a code location (vendored
contracts at tools/release/contracts/ and the runtime classes under
tools/release/src/) but is no longer a sub-package. tools/ is already
export-ignore'd in .gitattributes so this never shipped to users
either way.

Test counts: 52 isolated tests, 140 assertions (unchanged from the
sub-package layout). PHPStan level 10 + strict-rules clean across
the full 4310-file analysis. PHPCS clean. Composer-require-checker
clean.

Refs: openemr/openemr-devops#664
