+++ tests/Tests/Isolated/Release/BranchVersionResolverTest.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace OpenEMR\Release\Tests;
+namespace OpenEMR\Tests\Isolated\Release;
 
 use OpenEMR\Release\BranchVersionResolver;
 use PHPUnit\Framework\TestCase;
