+++ tests/Tests/Isolated/Release/TagVerifierTest.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace OpenEMR\Release\Tests;
+namespace OpenEMR\Tests\Isolated\Release;
 
 use OpenEMR\Release\TagVerifier;
 use PHPUnit\Framework\TestCase;
