--- tools/release/.gitignore
@@ -1,3 +1,0 @@
-/vendor/
-/.phpunit.cache/
-/.phpunit.result.cache
