--- tools/release/composer.json
@@ -1,52 +1,0 @@
-{
-  "name": "openemr/release-conductor",
-  "description": "Release-automation conductor tooling for openemr/openemr",
-  "type": "project",
-  "license": "GPL-3.0-or-later",
-  "require": {
-    "php": "^8.2",
-    "ext-json": "*",
-    "justinrainbow/json-schema": "^6.0",
-    "symfony/console": "^7.0",
-    "symfony/http-client": "^7.0",
-    "symfony/process": "^7.0"
-  },
-  "require-dev": {
-    "phpstan/phpstan": "^2.1",
-    "phpstan/phpstan-strict-rules": "^2.0",
-    "phpunit/phpunit": "^11.0",
-    "rector/rector": "^2.0",
-    "squizlabs/php_codesniffer": "^3.10"
-  },
-  "autoload": {
-    "psr-4": {
-      "OpenEMR\\Release\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "OpenEMR\\Release\\Tests\\": "tests/"
-    }
-  },
-  "config": {
-    "sort-packages": true
-  },
-  "scripts": {
-    "check": [
-      "@phpcs",
-      "@phpstan",
-      "@rector-check",
-      "@test"
-    ],
-    "fix": [
-      "@phpcbf",
-      "@rector-fix"
-    ],
-    "phpcbf": "phpcbf",
-    "phpcs": "phpcs",
-    "phpstan": "phpstan analyse --memory-limit=2G",
-    "rector-check": "rector process --dry-run",
-    "rector-fix": "rector process",
-    "test": "phpunit"
-  }
-}
