--- tools/release/phpcs.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<ruleset name="OpenEMR Release Conductor">
-    <description>Coding standard for openemr/openemr release-conductor tooling</description>
-
-    <arg name="extensions" value="php"/>
-    <arg name="colors"/>
-    <arg value="sp"/>
-
-    <file>src</file>
-    <file>bin</file>
-    <file>tests</file>
-
-    <exclude-pattern>*/vendor/*</exclude-pattern>
-
-    <rule ref="PSR12"/>
-
-    <rule ref="Generic.Arrays.DisallowLongArraySyntax"/>
-    <rule ref="Generic.PHP.DeprecatedFunctions"/>
-    <rule ref="Generic.Files.LineLength">
-        <properties>
-            <property name="lineLimit" value="120"/>
-            <property name="absoluteLineLimit" value="150"/>
-        </properties>
-    </rule>
-</ruleset>
