--- tools/release/phpstan.neon
@@ -1,14 +1,0 @@
-includes:
-    - vendor/phpstan/phpstan-strict-rules/rules.neon
-
-parameters:
-    level: 10
-    phpVersion: 80200
-    paths:
-        - src
-        - bin
-        - tests
-    checkMissingCallableSignature: true
-    checkUninitializedProperties: true
-    checkTooWideReturnTypesInProtectedAndPublicMethods: true
-    reportAlwaysTrueInLastCondition: true
