--- tools/release/phpunit.xml.dist
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:noNamespaceSchemaLocation="vendor/phpunit/phpunit/phpunit.xsd"
-         bootstrap="vendor/autoload.php"
-         colors="true"
-         failOnRisky="true"
-         failOnWarning="true"
-         cacheDirectory=".phpunit.cache">
-    <testsuites>
-        <testsuite name="release-conductor">
-            <directory>tests</directory>
-        </testsuite>
-    </testsuites>
-    <source>
-        <include>
-            <directory>src</directory>
-            <directory>bin</directory>
-        </include>
-    </source>
-</phpunit>
