--- tools/release/rector.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\ValueObject\PhpVersion;
-
-return RectorConfig::configure()
-    ->withPaths([
-        __DIR__ . '/src',
-        __DIR__ . '/bin',
-        __DIR__ . '/tests',
-    ])
-    ->withPhpVersion(PhpVersion::PHP_82)
-    ->withPreparedSets(
-        deadCode: true,
-        codeQuality: true,
-        typeDeclarations: true,
-    )
-    ->withCache('/tmp/rector');
